% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{dataset_query}
\alias{dataset_query}
\title{dataset_query function}
\usage{
dataset_query()
}
\value{
\code{data.frame} object
}
\description{
This function queries the "data.riksdagen" API and retrieves all the available information of Swedish politicians.
}
