% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{get_politicians_data}
\alias{get_politicians_data}
\title{get_politicians_data function}
\usage{
get_politicians_data(column_idx = c(5, 6, 11))
}
\arguments{
\item{column_idx}{is \code{numeric} vector that indicates the colums to select in the original data.}
}
\value{
a \code{data.frame}
}
\description{
This function gets the desired data of the Swedish politicians. It returns a \code{data.frame} containing the columns we want to extract from the original data,
plus an additional column with the years that politician has been working.
}
