% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{years_function}
\alias{years_function}
\title{years_function function}
\usage{
years_function(dataframe_var)
}
\arguments{
\item{dataframe_var}{is a \code{data.frame} object}
}
\value{
a \code{numeric} vector
}
\description{
Given a \code{dataframe} with the \code{from} and \code{tom} columns containing dates with format starting with the year,
this function returns a vector with all the years included in those intervals.
}
